Added Fully working Authentication System

+++ app/Http/Controllers/PostsController.php
@@ -7,6 +7,12 @@ use App\Post;
 
 class PostsController extends Controller
 {
+
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -45,6 +51,7 @@ class PostsController extends Controller
         $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        $post->user_id = auth()->user()->id;
         $post->save();
 
         return redirect('/posts')->with('success', 'Post Created');
@@ -71,6 +78,10 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+        // Check For Correct user
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error','Unauthorized Page');
+        }
         return view('posts.edit')->with('post',$post);
     }
 
@@ -105,6 +116,10 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        // Check for Current user
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error','Unauthorized Page');
+        }
         $post->delete();
         return redirect('/posts')->with('success', 'Post Deleted');
     }
